Add threaded connection logic
This commit adds ConnectionPool enum with a single instance of ConnectionPool that is lazily initialized
using an inline private static class. Also it adds a dummy Connection class with a sole purpose of tracking
IDs of would-be connections. In the Main class, we create a fixed thread pool of 5 threads as well as 2
separate threads, and we supply pooled threads with a longer version of connectionGet task. Then we start all
tasks at the same time so that separate threads release their connections earlier and pooled threads that are
waiting can get them.